Python/requirements.txt
Pillow==9.2.0
pycryptodome==3.15.0
PyPDF2==2.10.2
qrcode==7.3.1
rsa==4.9
zxing==0.14
